restructure, add prefix to card
