+++ polymorphia-frontend/components/course/event-section/RewardsInfo.tsx
@@ -1,5 +1,5 @@
 import ProgressBar from '@/components/progressbar/ProgressBar';
-import { RewardsInfoProps } from '@/interfaces/course/EventSectionInterfaces';
+import { RewardsInfoProps } from '@/interfaces/course/event-section/EventSectionInterfaces';
 import { API_STATIC_URL } from '@/services/api';
 import Image from 'next/image';
 import '../../../styles/event-section.css';
