--- polymorphia-frontend/interfaces/card/CardInterfaces.ts
@@ -1,11 +1,0 @@
-export interface CardProps {
-  id: number;
-  title: string;
-  subtitle?: string;
-  xp?: string;
-  onClick?: () => void;
-}
-
-export interface CardGridProps {
-  cards: CardProps[];
-}
