+++ polymorphia-frontend/interfaces/course/event-section/card/EventSectionCardInterfaces.ts
@@ -1,0 +1,11 @@
+export interface EventSectionCardProps {
+  id: number;
+  title: string;
+  subtitle?: string;
+  xp?: string;
+  onClick?: () => void;
+}
+
+export interface EventSectionCardGridProps {
+  cards: EventSectionCardProps[];
+}
