+++ polymorphia-frontend/styles/event-section-card.css
@@ -1,32 +1,32 @@
 @import 'tailwindcss';
 @import './globals.css';
 
-.card-grid {
+.event-section-card-grid {
   @apply w-full h-full grid grid-cols-1 gap-5;
   @apply md:grid-cols-2 md:max-w-2xl;
   @apply 2xl:grid-cols-3 2xl:max-w-5xl;
 }
 
-.card {
+.event-section-card {
   @apply w-full p-5 rounded-xl border-2 border-neutral-800 backdrop-blur-xs shadow-md transition-all h-fit;
 }
 
-.card-hover {
+.event-section-card-hover {
   @apply hover:shadow-lg hover:scale-[1.02];
 }
 
-.card h1 {
+.event-section-card h1 {
   @apply truncate text-5xl mb-2;
 }
 
-.card-subtitle {
+.event-section-card-subtitle {
   @apply truncate text-2xl;
 }
 
-.card-xp {
+.event-section-card-xp {
   @apply text-nowrap text-3xl;
 }
 
-.card-bottom {
+.event-section-card-bottom {
   @apply flex flex-row justify-between gap-3 items-end;
 }
